Refactoring disruption generation from history
